air-traffic/requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
backports.ssl==0.0.9
certifi==2017.4.17
cffi==1.10.0
cryptography==1.8.1
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pycparser==2.17
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
satori-sdk-python==1.0.3
six==1.10.0
